backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
djangorestframework-simplejwt